unsigned kinda work

+++ includes/ft_printf_internal.h
@@ -6,12 +6,13 @@
 /*   By: amarzial <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/12/08 17:49:26 by amarzial          #+#    #+#             */
-/*   Updated: 2016/12/09 20:00:22 by amarzial         ###   ########.fr       */
+/*   Updated: 2016/12/10 13:42:47 by amarzial         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_PRINTF_INTERNAL_H
 # include <stdarg.h>
+# include <stdint.h>
 # define FT_PRINTF_SIGNED_INTERNAL_H
 # define FT_PRINTF_H
 # define FT_PRINTF_FLAGS "#0- +"
@@ -47,7 +48,7 @@ void			ft_printf_handler(t_arg *arg, va_list *lst);
 
 int				ft_printf_signed_decimal(t_arg *arg, va_list *lst);
 
-char			*ft_printf_signedtostr(long long int num, t_arg *arg);
+char			*ft_printf_signedtostr(intmax_t num, t_arg *arg);
 
 void			ft_printf_putnchar(char c, int times);
